Extract PatternVisitor to easily patch it
This allow external users to extend escope with SpreadProperty support.
