+++ src/referencer.js
@@ -25,102 +25,9 @@ import { Syntax } from 'estraverse';
 import esrecurse from 'esrecurse';
 import Reference from './reference';
 import Variable from './variable';
+import PatternVisitor from './pattern-visitor';
 import { ParameterDefinition, Definition } from './definition';
 import assert from 'assert';
-
-class PatternVisitor extends esrecurse.Visitor {
-    constructor(rootPattern, callback) {
-        super();
-        this.rootPattern = rootPattern;
-        this.callback = callback;
-        this.assignments = [];
-        this.rightHandNodes = [];
-        this.restElements = [];
-    }
-
-    Identifier(pattern) {
-        const lastRestElement = getLast(this.restElements);
-        this.callback(pattern, {
-            topLevel: pattern === this.rootPattern,
-            rest: lastRestElement != null && lastRestElement.argument === pattern,
-            assignments: this.assignments
-        });
-    }
-
-    Property(property) {
-        // Computed property's key is a right hand node.
-        if (property.computed) {
-            this.rightHandNodes.push(property.key);
-        }
-
-        // If it's shorthand, its key is same as its value.
-        // If it's shorthand and has its default value, its key is same as its value.left (the value is AssignmentPattern).
-        // If it's not shorthand, the name of new variable is its value's.
-        this.visit(property.value);
-    }
-
-    ArrayPattern(pattern) {
-        var i, iz, element;
-        for (i = 0, iz = pattern.elements.length; i < iz; ++i) {
-            element = pattern.elements[i];
-            this.visit(element);
-        }
-    }
-
-    AssignmentPattern(pattern) {
-        this.assignments.push(pattern);
-        this.visit(pattern.left);
-        this.rightHandNodes.push(pattern.right);
-        this.assignments.pop();
-    }
-
-    RestElement(pattern) {
-        this.restElements.push(pattern);
-        this.visit(pattern.argument);
-        this.restElements.pop();
-    }
-
-    MemberExpression(node) {
-        // Computed property's key is a right hand node.
-        if (node.computed) {
-            this.rightHandNodes.push(node.property);
-        }
-        // the object is only read, write to its property.
-        this.rightHandNodes.push(node.object);
-    }
-
-    //
-    // ForInStatement.left and AssignmentExpression.left are LeftHandSideExpression.
-    // By spec, LeftHandSideExpression is Pattern or MemberExpression.
-    //   (see also: https://github.com/estree/estree/pull/20#issuecomment-74584758)
-    // But espree 2.0 and esprima 2.0 parse to ArrayExpression, ObjectExpression, etc...
-    //
-
-    SpreadElement(node) {
-        this.visit(node.argument);
-    }
-
-    ArrayExpression(node) {
-        node.elements.forEach(this.visit, this);
-    }
-
-    AssignmentExpression(node) {
-        this.assignments.push(node);
-        this.visit(node.left);
-        this.rightHandNodes.push(node.right);
-        this.assignments.pop();
-    }
-
-    CallExpression(node) {
-        // arguments are right hand nodes.
-        node.arguments.forEach(a => { this.rightHandNodes.push(a); });
-        this.visit(node.callee);
-    }
-}
-
-function getLast(xs) {
-    return xs[xs.length - 1] || null;
-}
 
 function traverseIdentifierInPattern(rootPattern, referencer, callback) {
     // Call the callback at left hand identifier nodes, and Collect right hand nodes.
@@ -131,18 +38,6 @@ function traverseIdentifierInPattern(rootPattern, referencer, callback) {
     if (referencer != null) {
         visitor.rightHandNodes.forEach(referencer.visit, referencer);
     }
-}
-
-function isPattern(node) {
-    var nodeType = node.type;
-    return (
-        nodeType === Syntax.Identifier ||
-        nodeType === Syntax.ObjectPattern ||
-        nodeType === Syntax.ArrayPattern ||
-        nodeType === Syntax.SpreadElement ||
-        nodeType === Syntax.RestElement ||
-        nodeType === Syntax.AssignmentPattern
-    );
 }
 
 // Importing ImportDeclaration.
@@ -442,7 +337,7 @@ export default class Referencer extends esrecurse.Visitor {
     }
 
     AssignmentExpression(node) {
-        if (isPattern(node.left)) {
+        if (PatternVisitor.isPattern(node.left)) {
             if (node.operator === '=') {
                 this.visitPattern(node.left, {processRightHandNodes: true}, (pattern, info) => {
                     var maybeImplicitGlobal = null;
@@ -506,7 +401,7 @@ export default class Referencer extends esrecurse.Visitor {
     }
 
     UpdateExpression(node) {
-        if (isPattern(node.argument)) {
+        if (PatternVisitor.isPattern(node.argument)) {
             this.currentScope().__referencing(node.argument, Reference.RW, null);
         } else {
             this.visitChildren(node);
